Fix packs after removal
The packIds property of each pack would be messed up after removing a
pack from the *middle* of the list. Now, decrementing those ids for any
packs that came after. It's not the most elegant way to do it (it runs
in linear time), but in practice there are only going to be around 15
items in the list at the very most.

Adding a pack: change default quantity. Setting it to 0 saves a call to
the expensive 'updateTotals' method.

Change qty for default pack (in constructor) to a power of 10. I feel
it makes it easier to 'get' upon first glance.